+++ chimera/render/VertexData.hpp
@@ -2,6 +2,7 @@
 #define __CHIMERA_VERTEX_HPP
 
 #include <glm/glm.hpp>
+#include <vector>
 
 // Tutorial opengl shadeGL - https://learnopengl.com/Getting-started/Shaders
 
@@ -17,5 +18,8 @@ struct VertexData {
     glm::vec3 normal;   // 3 * 4 = 12 (12 - 23)
     glm::vec2 texture;  // 2 * 4 = 08 (24 - 31)
 };
+
+void vertexDataIndexCompile(std::vector<VertexData>& inData, std::vector<VertexData>& outData, std::vector<unsigned int>& out_indices);
+
 } // namespace Chimera
 #endif
