+++ samples/simples/include/Game.hpp
@@ -37,7 +37,7 @@ class Game : public Chimera::IGameClientEvents {
     Chimera::TrackBall trackBall;
     Chimera::CanvasGL* pCanvas;
     Chimera::Shader* pShader;
-    Chimera::VertexRenderStatic VBO;
+    Chimera::VertexRenderStatic vertexRenderStatic;
     glm::mat4 projection;
     glm::mat4 view;
     glm::mat4 model;
