+++ samples/simples/src/Game.cpp
@@ -85,9 +85,9 @@ void Game::start() {
     Chimera::MeshData m;
     Chimera::LoaderObj loader;
     // loader.getMesh("./data/models/tela01.obj", m);
-    loader.getMesh("./data/models/square2.obj", m);
+    // loader.getMesh("./data/models/square2.obj", m);
     // loader.getMesh("./data/models/parede_simples.obj", m);
-    // loader.getMesh("./data/models/map02.obj", m);
+    loader.getMesh("./data/models/map02.obj", m);
 
     if (loader.hasMaterial() == true)
         loader.getMaterial(material);
@@ -105,7 +105,7 @@ void Game::start() {
     m.toVertexData(renderData);
 
     std::vector<unsigned int> index;
-    VBO.create(renderData, index);
+    vertexRenderStatic.create(renderData, index);
 }
 
 void Game::stop() {}
@@ -153,7 +153,7 @@ void Game::render() {
     // aplica a textura
     // pTex->apply(0, "material.tDiffuse", pShader);
 
-    VBO.render();
+    vertexRenderStatic.render();
 
     pCanvas->after();
     pCanvas->swapWindow();
